+++ examples/02_queries.ts
@@ -1,10 +1,10 @@
 import {
-  IncentivizedTestent,
+  IncentivizedTestnet,
   NibiruQueryClient,
   NibiruSigningClient,
 } from "@nibiruchain/nibijs"
 
-const TEST_CHAIN = IncentivizedTestent(1)
+const TEST_CHAIN = IncentivizedTestnet(2)
 
 async function runExample() {
   const queryClient = await NibiruQueryClient.connect(TEST_CHAIN.endptTm)