+++ packages/indexer-nibi/src/heart-monitor.test.ts
@@ -2,7 +2,7 @@ import { HeartMonitor } from "./heart-monitor"
 import { cleanResponse, gqlEndptFromTmRpc } from "./gql"
 
 const heartMonitor = new HeartMonitor({
-  endptTm: "https://hm-graphql.devnet-2.nibiru.fi",
+  endptTm: "https://hm-graphql.itn-2.nibiru.fi",
 })
 
 describe("Heart Monitor constructor", () => {
@@ -48,8 +48,8 @@ describe("gqlEndptFromTmRpc", () => {
 
   const tests: TestCase[] = [
     {
-      in: "https://rpc.devnet-2.nibiru.fi",
-      want: "https://hm-graphql.devnet-2.nibiru.fi/graphql",
+      in: "https://rpc.itn-2.nibiru.fi",
+      want: "https://hm-graphql.itn-2.nibiru.fi/graphql",
     },
     {
       in: "----rpc.itn-1.-----",
@@ -108,6 +108,29 @@ test("distributionCommissions", async () => {
     const fields = ["commission", "validator"]
     fields.forEach((field: string) => {
       expect(distributionCommissions).toHaveProperty(field)
+    })
+  }
+})
+
+test("governance", async () => {
+  const resp = await heartMonitor.governance({
+    govDeposits: {
+      limit: 1,
+    },
+    govProposals: {
+      limit: 1,
+    },
+    govVotes: {
+      limit: 1,
+    },
+  })
+  expect(resp).toHaveProperty("governance")
+
+  if (resp.governance) {
+    const governance = resp.governance!
+    const fields = ["govDeposits", "govProposals", "govVotes"]
+    fields.forEach((field: string) => {
+      expect(governance).toHaveProperty(field)
     })
   }
 })